Rename `loops/base.py` to `loops/loop.py` (#13043)
