+++ pl_examples/loop_examples/kfold.py
@@ -31,8 +31,8 @@ from pl_examples.basic_examples.mnist_datamodule import MNIST
 from pl_examples.basic_examples.mnist_examples.image_classifier_4_lightning_module import ImageClassifier
 from pytorch_lightning import LightningDataModule, seed_everything, Trainer
 from pytorch_lightning.core.module import LightningModule
-from pytorch_lightning.loops.base import Loop
 from pytorch_lightning.loops.fit_loop import FitLoop
+from pytorch_lightning.loops.loop import Loop
 from pytorch_lightning.trainer.states import TrainerFn
 
 #############################################################################################
