+++ pytorch_lightning/loops/__init__.py
@@ -12,10 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from pytorch_lightning.loops.base import Loop  # noqa: F401
 from pytorch_lightning.loops.batch import ManualOptimization  # noqa: F401
 from pytorch_lightning.loops.batch import TrainingBatchLoop  # noqa: F401
 from pytorch_lightning.loops.dataloader import DataLoaderLoop, EvaluationLoop, PredictionLoop  # noqa: F401
 from pytorch_lightning.loops.epoch import EvaluationEpochLoop, PredictionEpochLoop, TrainingEpochLoop  # noqa: F401
 from pytorch_lightning.loops.fit_loop import FitLoop  # noqa: F401
+from pytorch_lightning.loops.loop import Loop  # noqa: F401
 from pytorch_lightning.loops.optimization.optimizer_loop import OptimizerLoop  # noqa: F401
